requirements.txt
streamlit
transformers
torch
pandas
python-dotenv